Tera

+++ app/controllers/orders_controller.rb
@@ -47,19 +47,24 @@ class OrdersController < ApplicationController
   def create
     @order = Order.new(order_params)
     @order.customer_id = current_customer.id
-    @order.save
+    if @order.save
+      # 注文商品テーブル作成
+      current_customer.cart_items.each do |cart_item|
+        order_detail = @order.order_details.new
+        order_detail.order_id = @order.id
+        order_detail.item_id = cart_item.item_id
+        order_detail.price = cart_item.item.price
+        order_detail.amount = cart_item.amount
+        order_detail.save
+        cart_item.destroy
+      end
+      render :complete
 
-    # 注文商品テーブル作成
-    current_customer.cart_items.each do |cart_item|
-      order_detail = @order.order_details.new
-      order_detail.order_id = @order.id
-      order_detail.item_id = cart_item.item_id
-      order_detail.price = cart_item.item.price
-      order_detail.amount = cart_item.amount
-      order_detail.save
-      cart_item.destroy
+    else
+      @customer = current_customer
+      render :new
+
     end
-    render :complete
 
   end
